backend/requirements.txt
alembic
fastapi
httpx
openai
passlib[bcrypt]
psycopg2-binary
python-dotenv
python-jose[cryptography]
python-multipart
requests
safetensors>=0.4.0
sentence-transformers>=2.3.0
sqlalchemy
transformers>=4.36.0
uvicorn[standard]
